Setup folder utils
- Add file post.js
- Add file pagination.js
- Add file post.js
