+++ js/utils/index.js
@@ -1,2 +1,5 @@
 export * from './common'
 export * from './selectors'
+export * from './pagination'
+export * from './post'
+export * from './search'
